Take every item attribute from Backpack Battles
Our items were scraped from that game and renamed, then drifted. Shapes
were flattened to WxH rectangles, and costs, rarities and weapon numbers
were rewritten by hand until they matched nothing.

Each item now carries:

  map      the real grid, as a text map. '#' is a square the item covers,
           '*' and '+' are the star and diamond slots it reaches into.
           A Spike Launcher is a plus shape, not a 2x2.
  source   the item it came from, so the link survives the next rename
  sockets  how many gemstones it holds, though nothing sockets one yet
  icontype melee, ranged, magic and so on, which nothing reads yet

and its cost, rarity, damage, accuracy, cooldown and CPU cost are the
source item's. `uncommon` is gone, because Backpack Battles has six
rarities and that is not one of them.

`shape` stays until the map parser replaces it in config_loader.

CPU cost is a float now. Backpack Battles spends stamina in fractions and
we only ever spent whole numbers. The runtime was already fractional, so
only four annotations had to change.

The pool and its regeneration had to move with the costs, or a resource
nothing can exhaust would have stopped being a resource. Regeneration of
1 a second is theirs: the wiki calls the Mecha Bat's 0.5 a second "about
50% of the player's default stamina allowance", players report the same
figure in game, and the percentage buffs only work against a base of 1.
The pool of 3 is ours - they have never published theirs. Two weapons now
run level with regeneration and a third has to wait.

Battles started on 10 CPU whatever the pool held, which is how the old
value survived the change unnoticed. They start full instead.

+++ server/item_effects.py
@@ -190,7 +190,7 @@ class TimerTrigger(Trigger):
     """Activates on a timer"""
 
     cooldown: float
-    cpu_cost: int
+    cpu_cost: float
     effects: List[Effect] = field(default_factory=list)
 
     # Runtime state
@@ -224,7 +224,7 @@ class DamageTakenTrigger(Trigger):
 
     threshold: Optional[float] = None  # Only activate below X% health
     cooldown: float = 0.0  # Optional cooldown
-    cpu_cost: int = 0
+    cpu_cost: float = 0.0
     effects: List[Effect] = field(default_factory=list)
 
     # Runtime state
